1. Controller구조 변경  -web -> web, user, restaurant, board
