+++ SpringBootPractice-User/bin/DDL.txt
@@ -1,0 +1,23 @@
+CREATE DATABASE `User_MSA` /*!40100 DEFAULT CHARACTER SET utf8mb4 COLLATE utf8mb4_0900_ai_ci */;
+
+CREATE TABLE `User` (
+  `user_id` varchar(45) NOT NULL,
+  `user_password` varchar(45) DEFAULT NULL,
+  `user_name` varchar(45) DEFAULT NULL,
+  `user_email` varchar(45) DEFAULT NULL,
+  PRIMARY KEY (`user_id`)
+) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_0900_ai_ci;
+
+INSERT INTO `User_MSA`.`User`
+(`user_id`,`user_password`,`user_name`,`user_email`)
+VALUES('admin', 'admin', '손재하', 'jaeha@naver.com');
+
+INSERT INTO `User_MSA`.`User`
+(`user_id`,`user_password`,`user_name`,`user_email`)
+VALUES('seungho', 'seungho', '최승호', 'seungho93@naver.com');
+
+INSERT INTO `User_MSA`.`User`
+(`user_id`,`user_password`,`user_name`,`user_email`)
+VALUES('doming', 'dmongmin', '김동민', 'dongminrotc@naver.com');
+
+
