+++ SpringBootPractice-User/bin/README.md
@@ -1,0 +1,1 @@
+1. MSA를 위한 사용자 서비스 프로젝트
