+++ SpringBootPractice-User/bin/src/main/resources/application.properties
@@ -1,0 +1,8 @@
+server.port=8889
+spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
+spring.datasource.url=jdbc:mysql://localhost:3306/User_MSA?useUnicode=yes&characterEncoding=utf8&useSSL=false&serverTimezone=Asia/Seoul
+spring.datasource.username=root
+spring.datasource.password=12345
+spring.mvc.static-path-pattern=/static/**
+spring.resources.static-locations=classpath:/static/
+spring.resources.add-mappings=true
