+++ SpringBootPractice-User/bin/src/main/resources/mapper/UserMapper.xml
@@ -1,0 +1,32 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE mapper
+		PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
+		"http://mybatis.org/dtd/mybatis-3-mapper.dtd">
+		
+<mapper namespace="mapper.UserMapper">
+	<resultMap type="User" id="User">
+		<result property="user_id" column="user_id" javaType="java.lang.String" jdbcType="VARCHAR"/>
+		<result property="user_password" column="user_password" javaType="java.lang.String" jdbcType="VARCHAR"/>
+		<result property="user_name" column="user_name" javaType="java.lang.String" jdbcType="VARCHAR"/>
+		<result property="user_email" column="user_email" javaType="java.lang.String" jdbcType="VARCHAR"/>
+	</resultMap>
+	
+	<!-- 전체 사용자 조회 -->
+	<select id="selectAllUser" resultMap="User">
+	<![CDATA[
+		SELECT *
+		FROM User
+	]]>
+	</select>
+	
+	<!-- 사용자 가입  -->
+	<select id="createUser" parameterType="User">
+	<![CDATA[
+		INSERT INTO `User_MSA`.`User`
+		(`user_id`,`user_password`,`user_name`,`user_email`)
+		VALUES
+		(#{user_id}, #{user_password}, #{user_name}, #{user_email});
+	]]>
+	</select>
+	
+</mapper>
